Small Camera Change + Energy changes for Character States
Fact. The Rover is actually quite above the terrain(for those who don't
see it). Messing with it caused the Rover to fall through so I left it
alone. Scripts changed are the Camera, EnergyBar, and
ThirdPersonController so if you are having issues with gameLevel.unity,
you don't need to bring my changes to the scene; just the scripts are
enough.

I'll look into a rotation for the Rover later on today. Also, holding
the F key while not moving will orient the camera back to behind the
Rover, kinda like a reset feature. The Rover will not climb up steep
slopes; there is a check in ThirdPersonController.js with the
ControllerColliderHit to not jump upon hitting a surface with a certain
normal or above. Feel free to change the values of decayRates in the
Controller.js